+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 Changelog:
 
+- [2017-03-24] v0.1.3 - Experiment with viewport scroll sync. Seems not possible from extension, unfortunately.
+- [2017-03-24] v0.1.2 - Clears stale editors so that they can be closed without a save dialog appearing. Unfortunately the empty editors can be a bit confusing now as the helpful message is gone.
+- [2017-03-24] v0.1.1 - Reuses stale editors so that reopening a template won't always create a new untitled document.
+- [2017-03-24] v0.1.0 - Ctrl+Enter now uses another way to create untitled documents, which may work better on some platforms. Previous way of creating named documents is now Ctrl+Shift+Enter, use which one works for you.
+- [2017-03-24] v0.0.4 - Add SQL keyword to description
 - [2017-03-23] v0.0.3 - Add typescript as a dependency proper
 - [2017-03-23] v0.0.2 - Activate for js and ts
 - [2017-03-23] v0.0.1 - Initial proof-of-concept release